add db_path flag to binary and correct database columns
